Large refactor of the code
This commit modifies the organization of the project, hopefully helping
to extend the framework to new formulations with minimum overlap. The
storage classes are moved to the `SpatialDiscretization` folder, and the
`Mesh` class is also in its own folder.

The basic storage classes --`StateVector` and `FaceStateVector`-- are
modified to allow element-wise access regardless of the region. The
physical representation of the elements and faces is also different now,
being contained in large vectors accessed by views. This improves data
locality and makes data storage DOF-centered instead of
element-centered.

The monitor feature has its own place inside the project, ensuring that
new monitors can be added easily by simply writing a function and adding
them to the factory.

The dependencies with `StructArrays` and `DiffEqCallbacks` are dropped
also, since their functionality is more or less implemented in the code.