Use `Vector{SVector}` as the basic storage
Move from a matrix with dims `(ndofs, nvars)` to a vector of `ndofs`
`SVector`s of `nvars` components. It is all encapsulated in a new type
called `HybridArray`.

Limit `StdRegion`s to have the same number of nodes in all directions.
Although this feature might be interesting, it is definitely not useful
at this early stage of the development.