New module for approximation bases
The different node distributions and their associated bases are now in a
different module. I have also started implementing RBF stabilization and it is
available for the approximation of the solution (still need to implement this
for the flux approximation).

I also moved all FR related types and routines to the FR folder, and the more
generic functions of `IO.jl` and `Interfaces.jl` are now implemented in terms
of `AbstractSpatialDiscretization`. I should add a new intermediate abstract
type if I ever have other types of discretizations.