New storage layout (again)
All the work in Storage.jl goes to scratch/ with the new API. Data is
stored in plain arrays and the structures defined in Containers.jl
(`StateVector`...) and GlobalContainers.jl (`GlobalStateVector`...) allow
access to it through the methods: `dofs`, `vars`, `elements` and `faces`. The
underlying data can be accessed as `data`.